Moved core functionality to core section.
