+++ src/core/interfaces.ts
