+++ src/core/patterns.ts
