+++ src/core/validation.ts
