+++ src/operators/comparison.ts
@@ -1,6 +1,6 @@
-import {isZero, isNegative, isPositive} from '../validation';
+import {isZero, isNegative, isPositive} from '../core/validation';
 
-import {normalize, addLeftPadding, abs} from '../helpers';
+import {normalize, addLeftPadding, abs} from '../core/helpers';
 
 export const eq = (numOne: string, numTwo: string) =>
   normalize(numOne) === normalize(numTwo);
